+++ src/include.am
@@ -584,7 +584,8 @@ src_libwolfssl_la_SOURCES += \
                src/wolfio.c \
                src/keys.c \
                src/ssl.c \
-               src/tls.c
+               src/tls.c \
+               src/tlsx_base.c
 
 if BUILD_TLS13
 src_libwolfssl_la_SOURCES += src/tls13.c
