+++ src/tls.c
@@ -1309,135 +1309,6 @@ static WC_INLINE word16 TLSX_ToSemaphore(word16 type)
 #define TURN_OFF(semaphore, light) \
     ((semaphore)[(light) / 8] &= (byte) ~(0x01 << ((light) % 8)))
 
-/** Creates a new extension. */
-static TLSX* TLSX_New(TLSX_Type type, const void* data, void* heap)
-{
-    TLSX* extension = (TLSX*)XMALLOC(sizeof(TLSX), heap, DYNAMIC_TYPE_TLSX);
-
-    (void)heap;
-
-    if (extension) {
-        extension->type = type;
-        extension->data = (void*)data;
-        extension->resp = 0;
-        extension->next = NULL;
-    }
-
-    return extension;
-}
-
-/**
- * Creates a new extension and pushes it to the provided list.
- * Checks for duplicate extensions, keeps the newest.
- */
-static int TLSX_Push(TLSX** list, TLSX_Type type, const void* data, void* heap)
-{
-    TLSX* extension = TLSX_New(type, data, heap);
-
-    if (extension == NULL)
-        return MEMORY_E;
-
-    /* pushes the new extension on the list. */
-    extension->next = *list;
-    *list = extension;
-
-    /* remove duplicate extensions, there should be only one of each type. */
-    do {
-        if (extension->next && extension->next->type == type) {
-            TLSX *next = extension->next;
-
-            extension->next = next->next;
-            next->next = NULL;
-
-            TLSX_FreeAll(next, heap);
-
-            /* there is no way to occur more than
-             * two extensions of the same type.
-             */
-            break;
-        }
-    } while ((extension = extension->next));
-
-    return 0;
-}
-
-#ifdef WOLFSSL_TLS13
-
-/**
- * Creates a new extension and prepend it to the provided list.
- * Checks for duplicate extensions, keeps the newest.
- */
-static int TLSX_Prepend(TLSX** list, TLSX_Type type, void* data, void* heap)
-{
-    TLSX* extension = TLSX_New(type, data, heap);
-    TLSX* curr = *list;
-
-    if (extension == NULL)
-        return MEMORY_E;
-
-    /* remove duplicate extensions, there should be only one of each type. */
-    while (curr && curr->next) {
-        if (curr->next->type == type) {
-            TLSX *next = curr->next;
-
-            curr->next = next->next;
-            next->next = NULL;
-
-            TLSX_FreeAll(next, heap);
-        }
-        curr = curr->next;
-    }
-
-    if (curr)
-        curr->next = extension;
-    else
-        *list = extension;
-
-    return 0;
-}
-
-#endif /* WOLFSSL_TLS13 */
-
-#ifndef NO_WOLFSSL_CLIENT
-
-int TLSX_CheckUnsupportedExtension(WOLFSSL* ssl, TLSX_Type type);
-
-int TLSX_CheckUnsupportedExtension(WOLFSSL* ssl, TLSX_Type type)
-{
-    TLSX *extension = TLSX_Find(ssl->extensions, type);
-
-    if (!extension)
-        extension = TLSX_Find(ssl->ctx->extensions, type);
-
-    return extension == NULL;
-}
-
-int TLSX_HandleUnsupportedExtension(WOLFSSL* ssl);
-
-int TLSX_HandleUnsupportedExtension(WOLFSSL* ssl)
-{
-    SendAlert(ssl, alert_fatal, unsupported_extension);
-    return UNSUPPORTED_EXTENSION;
-}
-
-#else
-
-#define TLSX_CheckUnsupportedExtension(ssl, type) 0
-#define TLSX_HandleUnsupportedExtension(ssl) 0
-
-#endif
-
-/** Mark an extension to be sent back to the client. */
-void TLSX_SetResponse(WOLFSSL* ssl, TLSX_Type type);
-
-void TLSX_SetResponse(WOLFSSL* ssl, TLSX_Type type)
-{
-    TLSX *extension = TLSX_Find(ssl->extensions, type);
-
-    if (extension)
-        extension->resp = 1;
-}
-
 /******************************************************************************/
 /* Application-Layer Protocol Negotiation                                     */
 /******************************************************************************/
@@ -9308,35 +9179,6 @@ int TLSX_EarlyData_Use(WOLFSSL* ssl, word32 maxSz)
 /******************************************************************************/
 /* TLS Extensions Framework                                                   */
 /******************************************************************************/
-
-/** Finds an extension in the provided list. */
-TLSX* TLSX_Find(TLSX* list, TLSX_Type type)
-{
-    TLSX* extension = list;
-
-    while (extension && extension->type != type)
-        extension = extension->next;
-
-    return extension;
-}
-
-/** Remove an extension. */
-void TLSX_Remove(TLSX** list, TLSX_Type type, void* heap)
-{
-    TLSX* extension = *list;
-    TLSX** next = list;
-
-    while (extension && extension->type != type) {
-        next = &extension->next;
-        extension = extension->next;
-    }
-
-    if (extension) {
-        *next = extension->next;
-        extension->next = NULL;
-        TLSX_FreeAll(extension, heap);
-    }
-}
 
 /** Releases all extensions in the provided list. */
 void TLSX_FreeAll(TLSX* list, void* heap)
