+++ wolfssl.vcproj
@@ -176,6 +176,10 @@
 				>
 			</File>
 			<File
+				RelativePath=".\src\tlsx_base.c"
+				>
+			</File>
+			<File
 				RelativePath=".\src\tls13.c"
 				>
 			</File>
